NLP - Predict Movie Review Sentiment
The majority of this code was provided by Thomas Miller. This is the one assignment provided on github, where the template code makes up the majority of the code as this process is quite lengthly. I was required created two alternative RNN structures and performed hyperparameter testing. Using starter code as template, I then created four separate language models.